complete state feature

+++ src/PlayerManager.java
@@ -1,8 +1,8 @@
 public interface PlayerManager {
     public String getStateName();
     public void onEnterState();
-    public void takeDamage(int amount);
-    public void heal(int amount);
+    public void takeDamage(float amount);
+    public void heal(float amount);
     public void attack();
     public void kill();
 }
